requirements.txt
seleniumbase>=4.42.5,<5.0.0
selenium>=4.36.0,<5.0.0
pyautogui>=0.9.54,<0.10.0
# Note: python3-xlib is required for pyautogui on Linux/X11 (installed via apt-get in Dockerfile)
langgraph>=0.5.0,<0.6.0
langchain>=0.3.26,<0.4.0
langchain-google-genai>=2.1.6,<3.0.0
langchain-google-vertexai>=2.0.27,<3.0.0
langchain-core>=0.3.67,<0.4.0
langchain-community>=0.3.26,<0.4.0
python-dotenv>=1.1.1,<2.0.0
pydantic>=2.11.7,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0